pyproject.toml
[tool.poetry]
name = "pybites"
version = "0.1.0"
description = ""
authors = ["Malanius Privierre <malaniusprivierre@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
beautifulsoup4 = "^4.11.1"
requests = "^2.28.1"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = {version = "^22.3.0", allow-prereleases = true}
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
